backend/pyproject.toml
[project]
name = "fusionems-core"
version = "0.1.0"
description = "FusionEMS Enterprise Core backend"
requires-python = ">=3.11"
dependencies = [
  "fastapi>=0.110.0",
  "uvicorn>=0.30.0",
  "sqlalchemy>=2.0.30",
  "psycopg[binary]>=3.1.19",
  "alembic>=1.13.1",
  "pydantic-settings>=2.2.1",
  "python-jose[cryptography]>=3.3.0",
  "passlib[bcrypt]>=1.7.4",
  "email-validator>=2.1.1",
  "redis>=5.0.0",
  "requests>=2.32.0",
  "paramiko>=3.5.0",
  "boto3>=1.34.0",
  "stripe>=8.0.0",
  "telnyx>=2.0.0",
  "python-multipart>=0.0.9",
  "tenacity>=8.2.3",
  "openai>=1.30.0",
  "prometheus-client>=0.20.0",
  "opentelemetry-api>=1.26.0",
  "opentelemetry-sdk>=1.26.0",
  "opentelemetry-exporter-otlp>=1.26.0",
  "opentelemetry-instrumentation-fastapi>=0.47b0",
  "opentelemetry-instrumentation-requests>=0.47b0",
]

[tool.pylint.main]
init-hook = 'import sys; sys.path.insert(0, ".")'

[tool.pytest.ini_options]
pythonpath = ["."]
testpaths = ["tests"]
asyncio_mode = "auto"
markers = ["smoke: Smoke tests that require a live database and AWS credentials"]

[tool.ruff]
target-version = "py311"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "W", "I", "UP", "B", "SIM"]
ignore = [
  "E501",
  "B008",
  "B904",
]

[tool.ruff.lint.per-file-ignores]
"alembic/versions/*.py" = ["E501", "UP", "I001", "SIM"]
"tests/*.py" = ["E501", "E402"]
"core_app/models/__init__.py" = ["F401"]
"core_app/main.py" = ["E402"]
"core_app/epcr/validation_engine.py" = ["SIM102"]

[tool.mypy]
python_version = "3.11"
ignore_missing_imports = true
warn_unused_ignores = true
